Django-Rest-Framework/requirements.txt
asgiref==3.8.1
asttokens==3.0.0
certifi==2024.12.14
charset-normalizer==3.4.1
decorator==5.1.1
dj-rest-auth==7.0.1
Django==5.1.4
django-allauth==65.3.1
django-extensions==3.2.3
django-rest-auth==0.9.5
djangorestframework==3.15.2
executing==2.1.0
Faker==33.3.1
idna==3.10
ipython==8.31.0
jedi==0.19.2
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
